LayoutTests/platform/chromium-mac/fast/dom/Element/scale-page-bounding-client-rect-expected.txt

CONSOLE MESSAGE: line 12: Uncaught TypeError: Object [object Object] has no method 'scalePageBy'
This test must be run via DumpRenderTree.
